requirements.txt
asgiref==3.7.2
attrs==23.1.0
binary==1.0.0
contourpy==1.2.0
cycler==0.12.1
dataclasses==0.6
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-heroku==0.3.1
djangorestframework==3.14.0
djongo==1.2.31
dnspython==2.4.2
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
et-xmlfile==1.1.0
fonttools==4.47.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
patsy==0.5.4
Pillow==10.1.0
psycopg2==2.9.9
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.6.0
